requirements.txt
ollama==0.3.3
openai==1.52.2
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.5.2
